feat: Refactoring Application event code.

+++ structura-lib/src/geometry/mod.rs
@@ -2,10 +2,10 @@
 //!
 //!
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Copy, Clone)]
 pub struct Point {
-    pub x: f32,
-    pub y: f32,
+    pub x: f64,
+    pub y: f64,
 }
 
 #[derive(Debug, Clone)]
